Add a parser for DPX headers
largely carried over from Depix. Since we need a limited amount
of information - mainly resolution - we package all the binary
fields from Depix into structs where most of the members
are omitted (replaced with `x`). We read the entire chunk
in one go and use the sizing specifiers to determine how much
we need to read and how much we need to skip