Allow parsers to register themselves
so that we do not have to add explicit requires, and so that
an application using this library can add it's own parser
implementations/modules into the mix.

We keep the usage of .new in the parse method because the
method has to be reentrant - we have to ensure that it can
be called at the same time from multiple threads.