Add a custom exception for skips/reads
This exception would mean that we do not
have enough data in the IO regardless of
how far the parser could get, and we are unable
to finish parsing. This is an exception we need
to rescue - in the future we might want to
surround this with a catch() instead since
once we are at a point when the data cannot
be summoned we should let the parser bail out
cleanly. And exceptions are not the best for
flow control.