Get rid of "packed" notation for properties
We used it because you could specify a single
packspec (pattern to pass to "String#unpack")
by joining all the even elements of the array
together. In practice, we will do straight-ahead
reads which makes the parser code more readable

Should also prove useful in light of Rubocop rules incoming.