Use Hashes for MOOV atom data fields, remove repack() (#59)
* Get rid of "packed" notation for properties

We used it because you could specify a single
packspec (pattern to pass to "String#unpack")
by joining all the even elements of the array
together. In practice, we will do straight-ahead
reads which makes the parser code more readable

Should also prove useful in light of Rubocop rules incoming.

* Remove 0-padding

We read 4 bytes and we should be also having 4 bytes

* Correct the atom size calculation

There are more things to account for. We also need
tests for the Decoder, pronto