Use register_parser_constructor for natures/fmts
So @tehAnswer had a point here - we can simplify the code
quite a bit if we preselect parsers matching the natures/formats
we expect. However, there is an additional benefit. Currently, we
effectively require something registered as a parser to support the
following API:

  #natures
  #formats
  #call(io)
  .new

In the new setup we bring this down to one method:

  #call(io)

How do we do this? Well, first we move the two
methods you would have to manually define into the
call to `register_parser`. When we register a parser,
we pass either a callable or a class, and the module
figures out whether it has to call #new on the thing
given to it. This spares us from having  to have `.new`
defined - and thus spares us having to have parsers
be a Class in the first place. We can make a parser that
is a Proc instead, as long as we register it having
the right natures/formats.

Second, we supply the natures and the formats as the arguments
to the `register_parser` method - we tell FormatParser that
when someone wants to have anything of the kind of `:image`,
our parser is one that can provide that. We don't need to
supply that data from the parser object necessarily, and when
I asked to have this I was actually mistaken - because any
interface beyond just something callable is in this case
redundant.

So now, when we do the parser registration, we can do this

   FormatParser.register_parser self, natures: :image, formats: :jpg

and the module will figure out that it has to call .new on us
if we respond to it, and if we don't - the we will be considered
callable as is