Move the stdlib require to the top
We put require_relative within the module
so that once we have declared a module
named Foo we can then immediately require
a file containing `class Foo::Bar` and have
it know that it becomes a constant within
that module. This way we do not have to
use all the extra indentation in every file
we require.