Fix ZIP entry filename parsing
When ZIPs get created without the proper EFS flag in the general
purpose flags bitfield and the filenames include upper-ASCII
characters all bets are off, since these filenames are in the local
flavor of the DOS codepage chosen on the user's OS - usually Windows.

When we encounter those filenames we have no other option than to use
a Unicode replacement character since it is unlikely we will be able
to recover the correct filename without knowing what the locale was
at the time of archive creation. You can try to solve this problem
using word dictionaries per language, and using some text encoding
guesstimator to handle these cases, but for our purposes it is enough
that we can provide reasonable output that will be usable in general
UTF-8 contexts.