Fix EXIF parsing in JPEG files
When parsing JPEGs via EXIFR what EXIFR would do is actually
the following: it would seek to the start of the JPEG, scan
each marker (just like we do), scan the SOF markers
(just like we do), then find the APP1 marker (just like we do)
and then try to read that and parse the contents of the APP1
marker as if it came from a TIFF file.

This, in turn, would cause us to perform at least twice
as many read() calls on our IO as necessary, which triggers the
read limiter in some edge cases since there would be so many.
We also can't use the opportunity to read all the EXIF data from
upstream using a single read() - which, APP1 being a single unit
of data within the file - should be perfectly possible.

We remove all the JPEG-specific stuff from our EXIFR adapter
and pass it the "TIFF-ish" contents of the APP1 marker to deal
with, which should remove that problem.

The fact that we encountered so many problems with JPEGs
causing too many reads can be directly attributed to this
misuse of EXIFR.