Reject Keynote documents in JPEG parser
Keynote documents start with a perfectly sane sequence
of JPEG markers, including EXIF and all, probably for
Quicklook or for icon previews on macOS. They are, however,
ZIP files under the hood - and we want to keep them for
later for proper identification underneath the ZIP parser
that we already have. Also, if you rename a .key file
into a .jpg no image editor will still open that file,
so we can't consider this an "ambiguous detection".

In theory it could be possible to "trim" the ZIP data
from the Keynote file and still try to open it as if
it were a JPEG, but this is clearly out of scope for
what we are trying to do here.

Also replaces the truncated Keynote file we used previously
with a complete Keynote file created by us.