Speed up JPEG parser for invalid inputs (#105)
Previously we could keep scanning JPEG files until reaching the read limits,
which was also slow because of the incurred read/unpack/first calls for each
byte. This could lead to parse calls for, say, PSD files take 3-4 seconds since
we would be looking for a marker and not finding it and would just keep reading
one byte at a time.

This will make read_next_marker fail earlier if it cannot find a marker sequence
after "blind scanning" 1024 bytes.