Do not read all APP1 markers looking for EXIF
When we do our scanning we would previously read
the entire marker contents for the APP1 marker and then
try to determine whether it was EXIF or not. This strategy
was actually not very good since there are apparently JPEGs
where there is a ton of APP1 markers at the beginning,
none of which are EXIF to begin with. We will discard
those markers once we have ascertained that their header -
the first N bytes - are not matching the Exif magic string