Add a binary called format_parser_inspect
which allows you to get output from format_parser without
doing it from code. This is primarily useful when debugging
format detection issues, but can also be used as a shortcut
to determine what the file actually is. Think about it
as $file on steroids.

It will print detection information on STDOUT