Fix JPEG parsing regression
In ac9fbbbd6fed174027c750a0bf5be3e0ee49b05c we introduced
an optimization when looking for the next marker. It was
implemented using "reads in blocks of 2 bytes", which is
actually a bug since we are looking for a _sequence_
of bytes, and that sequence may end up smashed between
two of these blocks. Replace this way of reading with
reads "in groups of 2 byes" where you can examine the
tuple of the two bytes individually.