Fix TIFF parser regression
When we introduced CR2 (or at some other point) we created
a regression in the TIFF parser and it started reading the image
dimensions from all the wrong places. This resulted in some TIFFs
being recognized as having the image width/height of 1, which
was quite unfortunate.

Knowing that we have chosen, willingly, to have EXIFR at
our disposal, we will delegate parsing to it - it does
it fairly sanely, skips over the image content and reads
the minimum amount of data needed. And the best code is
code you don't have.

It also demonstrates, yet again, that asserting that
"this parser returns an Integer for width" is not
sufficient to make claims of any order, and we
probably should pay attention more when writing tests.
This is the third time we are encountering the pattern.

Also: yes, it's a TIFF.