JPEG: Skip over APP1 markers which are too small
If there is certainly not enough data in an APP1
to begin with, return early and avoid reading
anything from APP1 at all. This also avoids a case
where we would try to read into the EOF if the
APP1 marker is too tiny and is located near
the end of file - we would try to read it anyway.