Add a binary called `format_parser_inspect` (#107)
which allows you to get output from format_parser without
doing it from code. This is primarily useful when debugging
format detection issues, but can also be used as a shortcut
to determine what the file actually is. Think about it
as $file on steroids.

It will print detection information on STDOUT

Also adds Add FormatParser.parse_file_at which is
a convenience that can be very useful at times.
Because how often do you want to be doing File.open?