Make the reads limiter configurable
To make it possible to tweak performance on a per-app
basis we need to be able to play with things like cache
page size without releasing/testing a new version of the
gem every time. We make the limiter configuration
an explicit dependency - it can be given on each
call to `parse`, `parse_http` and friends.

Also restructures the main spec file for better
readability (and less nesting)