Make sure JSON output strings are sanitized (#114)
Used for sanitizing values that are sourced to `JSON::Generator::State#generate`
The reason we need to do this is as follows: `JSON.generate / JSON.dump / JSON.pretty_generate`
use a totally different code path than `"foo".to_json(generator_state)`. We cannot predict
which one of these two ways our users will be using, and at the same time we need to prevent
invalid Strings (ones which cannot be encoded into UTF-8) as well as Float::INFINITY values
from being passed to the JSON encoder. Since we cannot override the JSON generator with
these additions, instead we will deep-convert the entire object being output to make sure
it is up to snuff.