Fix the the thousand ways our ID3V2 parser is broken by using a library (#115)
Replace homegrown ID3 parser with id3tag.

id3tag is a mature library, it has version constraints similar to ours,
and it is compatible with our IO model. We could of course continue
torturing ourselves with a custom parser, but for this - just as is the case
for EXIF - using a third-party library makes sense. We keep it isolated from
our IO stack by passing it prepackaged buffers.

Protect from reading ID3 tags which are too large.