Remove ZIP checks for JPEGs
since these Keynote files actually start with a ZIP
local header, as they should - it's just that we
used to _scan_ until the SOI marker was found, which
the JPEG parser would do just right. Now we check
that the file starts with that marker, and this check
can go. We can also avoid reading an extra page of
content thanks to that.