Clean up the Ogg parser a little bit
Remove recursion when looking for the granule position. Also be more
explicit when doing lookups and string slices, and return early
if any preconditions are not satisfied. Use the same index
lookup we use in the ZIP parser for finding all locations of Ogg -
recursion can seem nicer, but with recursion you need to mention
offsets all over the place, which is not ideal - and makes it
easy to make a mistake.

Specify endianness explicitly for all `unpack` calls, since we
do not know which endianness the code will run on in practice.