Add :priority when registering parsers, use it when parsing (#132)
Allow the parser priority to be specified. The parsers with a lower priority value will be applied first.
We need this because of two important factors:

Some parsers are prone to misdetection. For instance, MP3 can be detected in files that aren't really MP3s, just by virtue of having "sort of" a number of MPEG frames. Other parsers, in contrast, are very quick to return early or to detect correctly. Since we rely on "return first" most often, it makes sense to apply the "safer" (most certain) parsers first
Our file format popularity is roughly this: JPEG, PDF, PNG, everything else. It makes sense to apply parsers for the more popular formats first, since a detection is likely to happen earlier.