Make sure we can "splice" multiple EXIF frames
JPEG files can contain multiple EXIF blocks,
one EXIF block per APP1 marker. We need
to "blend" them and preserve (and use) the
EXIF values only from the last block that
specifies those values. The values of the
EXIF tags that come later in the file should
overwrite the values that come from a block
that occurs earlier.

Additionally, while at it, we need to treat
"orientation" in a special way because it
can specify 0 for "unknown", in which case
we need to skip it and let the earlier value
be used instead.