Add support for MPEG video files (#149)
 Returns the following information:

- format type
- width px
- height px
- aspect ratio
- frame rate

Closes https://github.com/WeTransfer/format_parser/issues/138

## Issues found

Tests run okay, but if I try running an example script, seems like the parser is not properly registered. Might be a file loading issue when testing the gem locally?

Calculate file size and duration. I wasn't able to parse this properly with the information I found online

## Examples found

Different test examples from this website: https://standaloneinstaller.com/blog/big-list-of-sample-videos-for-testers-124.html. The license of the website is not clear and uses videos from different sources. Should I get files from another source with a proper license?