Do not return an Integer from MPEGParser#call
If the .times iterator has finished without returning early,
the .call method return value (number of times iterated) will
become the implicit return value for .call, returning not a `nil`
and not a `Result` but an `Integer`, which breaks the contract.

Implicit returns are OK but we have to be aware of them.