Add the ability to plug FormatParser intoto ActiveStorage (#158)
We want to be able to easily parse any file using format_parser when
using ActiveStorage in Rails apps.

To enable FormatParser in Rails, add it to the list of ActiveStorage
analyzers in your Rails config:

    config.active_storage.analyzers.prepend FormatParser::ActiveStorage::BlobAnalyzer

For the feature to work we wrap any ActiveStorage blob with an IO-like
object which then gets processed by FormatParser.

To test this new integration, we use a script from the ActiveStorage guide,
using RSpec for tests. The ruby script
can be run standalone:

- `ruby spec/integration/active_storage/rails_app.rb`

To test against S3 you need to change the active_storage.service and
active_storage.service_configuration values. But FormatParser should be compatible with
any ActiveStorage service.