Fix mp3_parser to not raise error when a tag frame has an unsupported encoding (#161)
We found a file which raises an error when calling the method tag.album (in id3v2), so I added a loop to call the member methods in the tag inside attempt_id3_v2_extraction. If some of them raise an error, it will consider the tag as invalid and return nil.

Since quite a few files have this problem we added a fallback encoding to the id3tag gem, and we configure it in format_parser when parsing a file. This should substantially reduce the number of errors and allow us to recover more data from MP3s (and most importantly - from more MP3 files).

Closes #160