Fix FormatParser.parse with results: :first to be deterministic (#184)
We were getting different results from `FormatParser.parse` using the option 
`results: :first`, which is the default value.

That resulted in flaky tests and in different results in production.

The reason for that was the method `parsers_for`, that used to sort the parsers 
only by them priorities, and when they were the same, Ruby didn't guarantee a 
stable result (see https://ruby-doc.org/core-2.5.0/Array.html#method-i-sort)

To fix that, we changed the class variable @parsers, which was a Set, to be an Array,
then we can use the order that the parsers were registered as the second criteria
to sort them. With that, `results: first` became deterministic.