Fix small and empty HTTP response handling
If a response is very small (or the HTTP range requested "rolls off" the end
of the resource body) S3 will respond with the full resource and with a 200 status code.
If a resource is empty no Range support will be available (and there is no good way to
address a 0-size resource using the HTTP Ranges specification).

To handle this better we will now allow small 200 responses to be treated as normal responses,
as long as they do not exceed the range size we request from the server.